Refactor cron job handlers to use separate handler files
Otherwise they are interdependent and all need chromium
