+++ backend/checkStaleStories.ts
@@ -1,0 +1,4 @@
+import checkStaleStoriesImpl from './src/cron/checkStaleStories';
+import withSetup from './withSetup';
+
+export const handler = withSetup(checkStaleStoriesImpl);
