+++ backend/generateStoryTitles.ts
@@ -1,0 +1,5 @@
+import generateStoryTitlesImpl from './src/cron/generateStoryTitles';
+
+import withSetup from './withSetup';
+
+export const handler = withSetup(generateStoryTitlesImpl);
