+++ backend/registerPrintfulWebhooks.ts
@@ -1,0 +1,4 @@
+import registerWebhooks from './src/business/merch/registerWebhooks';
+import withSetup from './withSetup';
+
+export const handler = withSetup(registerWebhooks);
